DTSAM-1278 Move common code into BaseProviderTest and update to spring6
